[COMMON] FIX: 온라인 유저 정보 중복되는 문제 해결

+++ backend/src/game/game.gateway.ts
@@ -333,19 +333,28 @@ export class GameGateway implements OnGatewayDisconnect {
 
     // 15초 후 초대장이 남아있다면 초대장을 삭제
     setTimeout(() => {
-      this.gameRoomService.deleteInvitationById(invitation);
-      const inviterSocketInfo =
-        this.statusService.getUserSocketInfoByUserId(userId);
+      console.log(
+        this.gameRoomService.findInvitationByInviteeId(
+          inviteeSocketInfo.userId,
+        ),
+      );
+      if (
+        this.gameRoomService.findInvitationByInviteeId(inviteeSocketInfo.userId)
+      ) {
+        this.gameRoomService.deleteInvitationById(invitation);
+        const inviterSocketInfo =
+          this.statusService.getUserSocketInfoByUserId(userId);
 
-      // 초대한 사람에게 시간초과 알림
-      this.server
-        .in(inviterSocketInfo.gameSocketId)
-        .emit('rejectInviteGameRoom');
+        // 초대한 사람에게 시간초과 알림
+        this.server
+          .in(inviterSocketInfo.gameSocketId)
+          .emit('rejectInviteGameRoom');
 
-      // 초대받은 사람에게 시간초과 알림
-      this.server
-        .in(inviteeSocketInfo.gameSocketId)
-        .emit('timeoutInviteGameRoom');
+        // 초대받은 사람에게 시간초과 알림
+        this.server
+          .in(inviteeSocketInfo.gameSocketId)
+          .emit('timeoutInviteGameRoom');
+      }
     }, 1000 * 15);
   }
